+++ src/core/aperture.h
@@ -25,7 +25,7 @@ public:
     // the final output is resized to the given dimensions
     image get_cfft(const image& aperture, const glm::ivec2& dims);
 
-    image get_flare(const image& cfft, float radius);
+    image get_flare(const image& cfft, int radius);
 
 private:
     aperture& operator=(const aperture& other);
